requirements.txt
boto3==1.12.0
botocore==1.15.0
Django==2.2.6
django-storages==1.9.1
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
Pillow==6.2.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
